+++ src/styles/slick.css
@@ -35,4 +35,3 @@
 .slick-slide > div > div > div {
 	width: 100%;
 }
-
